# LAB/LAB_04/server/api/openapi.yaml
openapi: 3.0.3
info:
  title: Film Manager API
  description: |-
    REST APIs for the interaction with the Film Manager service.
  termsOfService: http://swagger.io/terms/
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
...
servers:
  - url: http://localhost:3001/
    description: "Local Development Server"
...
paths:
...
  /api/films/public/{filmId}/images:
    get:
      tags:
        - api/films/public/{filmId}/images
      summary: Retrieve a list of images for a public film
      description: This operation retrieves a list of image structures associated with the public film with ID `filmId`. The response contains metadata for each image, such as its ID, type, and other relevant details, but not the actual image content itself. This operation does not require authentication.
      operationId: getImageListForPublicFilm
      parameters:
        - name: filmId
          in: path
          description: ID of the film to retrieve the images
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Images'
              examples:
                imageListExample:
                  summary: Example list of images for a public film with ID 1
                  value:
                    - id: 1
                      originalname: "poster"
                      filename: "image-123"
                      filmId: 1
                      self: "api/films/public/1/images/1"
                      delete: "api/films/public/1/images/1"
                    - id: 2
                      originalname: "clock"
                      filename: "image-456"
                      filmId: 1
                      self: "api/films/public/1/images/2"
                      delete: "api/films/public/1/images/2"
                    - id: 3
                      originalname: "tv"
                      filename: "image-789"
                      filmId: 1
                      self: "api/films/public/1/images/3"
                      delete: "api/films/public/1/images/3"
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                default:
                  summary: "Resource not found"
                  value:
                    code: 404
                    message: "the requested resource could not be found."
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                default:
                  summary: "Unexpected error occurred"
                  value:
                    code: 500
                    message: "an unexpected error occurred on the server. Please try again later."
    post:
      tags:
        - api/films/public/{filmId}/images
      security:
        - cookieAuth: []
      summary: "Associate a new image to a public film"
      description: "The image sent in request body is associated to the public film characterized by the ID specified in the path. Only the film owner can associate an image to the film."
      operationId: addImage
      parameters:
        - name: filmId
          in: path
          description: ID of the film
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: "The image to be assigned to the film"
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                image:
                  type: string
                  format: binary
                  description: "The image file to be uploaded (PNG, JPG, GIF formats supported)."
            encoding:
              image:
                contentType: image/png, image/jpg, image/jpeg, image/gif
        required: true
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Image'
              examples:
                Image Created Successfully:
                  value:
                    id: 1
                    originalname: "poster"
                    filename: "image-123"
                    filmId: 1
                    self: "api/films/public/1/images/1"
                    delete: "api/films/public/1/images/1"
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                default:
                  summary: "Authentication is required or failed"
                  value:
                    code: 401
                    message: "authentication is required. Please log in with valid credentials."
        '403':
          description: Forbidden (user is not the owner)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                default:
                  summary: "Access is forbidden"
                  value:
                    code: 403
                    message: "you do not have permission to access this resource."
        '404':
          description: Not Found (e.g., film with this id is not found)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                default:
                  summary: "Resource not found"
                  value:
                    code: 404
                    message: "the requested resource could not be found."
        '415':
          description: Unsupported Media Type (e.g., the request payload format is not supported)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                default:
                  summary: "Unsupported media type"
                  value:
                    code: 415
                    message: "The media type of the request payload is not supported by the server."
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                default:
                  summary: "Unexpected error occurred"
                  value:
                    code: 500
                    message: "an unexpected error occurred on the server. Please try again later."
    delete:
      tags:
        - api/films/public/{filmId}/images
      security:
        - cookieAuth: []
      summary: Delete an image associated to a public film
      description:  |
        All the images associated to the film with ID `filmId` are removed. Only the film owner can delete the images.
      operationId: deleteAllImagesAboutFilm
      parameters:
        - name: filmId
          in: path
          description: ID of the film
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: No Content
          content: {}
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                default:
                  summary: "Authentication is required or failed"
                  value:
                    code: 401
                    message: "authentication is required. Please log in with valid credentials."
        '403':
          description: Forbidden (user is not the owner)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                default:
                  summary: "Access is forbidden"
                  value:
                    code: 403
                    message: "you do not have permission to access this resource."
        '404':
          description: Not Found (e.g., film with this id is not found)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                default:
                  summary: "Resource not found"
                  value:
                    code: 404
                    message: "the requested resource could not be found."
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                default:
                  summary: "Unexpected error occurred"
                  value:
                    code: 500
                    message: "an unexpected error occurred on the server. Please try again later."
...
  /api/users:
    get:
      tags:
        - api/users
      security:
        - cookieAuth: []
      summary: Get information about the users
      description: |
        The available information (passwords excluded) about all the users is retrieved. This operation is available only to authenticated users. A pagination mechanism is used to limit the size of users.
      operationId: getUsers
      parameters:
        - in: query
          name: pageNo
          required: false
          schema:
            type: integer
          description: The id of the requested page (if absent, the first page is returned)
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Users'
              examples:
                users_page_1:
                  value:
                    totalPages: 2
                    currentPage: 1
                    totalItems: 2
                    users:
                      - id: 1
                        email: "user1@example.com"
                        name: "John Doe"
                        self: "/api/users/1"
                      - id: 2
                        email: "user2@example.com"
                        name: "Jane Doe"
                        self: "/api/users/2"
                    next: "/api/users?pageNo=2"
                    previous: null
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                default:
                  summary: "Authentication is required or failed"
                  value:
                    code: 401
                    message: "authentication is required. Please log in with valid credentials."
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                default:
                  summary: "Unexpected error occurred"
                  value:
                    code: 500
                    message: "an unexpected error occurred on the server. Please try again later."
  /api/users/{userId}:
    get:
      tags:
        - api/users/{userId}
      security:
        - cookieAuth: []
      summary: Get information about a user
      description: |
        The available information (password excluded) about the user specified by `userId` is retrieved. This operation requires authentication.
      operationId: getSingleUser
      parameters:
        - name: userId
          in: path
          description: ID of the user to get
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
              examples:
                user_found:
                  value:
                    id: 1
                    email: "user1@example.com"
                    name: "John Doe"
                    self: "/api/users/1"
        '400':
          description: Bad Request (Invalid user ID)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                default:
                  summary: "Invalid user ID"
                  value:
                    code: 400
                    message: "Invalid user ID. Please provide a valid numeric ID."
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                default:
                  summary: "Authentication is required or failed"
                  value:
                    code: 401
                    message: "authentication is required. Please log in with valid credentials."
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                default:
                  summary: "Resource not found"
                  value:
                    code: 404
                    message: "the requested resource could not be found."
  /api/users/authenticator:
    post:
      tags:
        - api/users/authenticator
      summary: Logs a user in
      description: |
        The user who wants to log in sends the user data to the authenticator which performs the operation.
        If the request for the login of a new user comes from an already authenticated user, the previous user is first logged out.
      operationId: authenticateUser
      requestBody:
        description: The data of the user who wants to perform log in. The data structure must contain email and password.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewUser'
            examples:
              validUserLogin:
                summary: Login request for an existing user with correct credentials
                value:
                  email: "john.doe@example.com"
                  password: "securePassword123"
              invalidPasswordLogin:
                summary: Login attempt with incorrect password
                value:
                  email: "jane.smith@example.com"
                  password: "wrongPassword987"
              newUserLogin:
                summary: Login for a new user registering with a valid email and password
                value:
                  email: "newuser@example.com"
                  password: "newUserPassword2024"
        required: true
      responses:
        '204':
          description: No Content (login successful)
          headers:
            Set-Cookie:
              schema:
                type: string
                example: "token=abcde12345; Path=/; HttpOnly"
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                default:
                  summary: "Generic bad request error"
                  value:
                    code: 400
                    message: "the request is malformed or invalid. Please check the request data."
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                default:
                  summary: "Authentication is required or failed"
                  value:
                    code: 401
                    message: "authentication is required. Please log in with valid credentials."
components:
  schemas:
...
    UserBase:
      description: "base schema for user-related information, including email and username."
      type: object
      additionalProperties: true
      properties:
        $schema:
          type: string
          description: "path to the JSON Schema to use for the validation of the JSON data structure describing the user"
        email:
          type: string
          format: email
          description: "user email"
        name:
          type: string
          description: "username"
      required:
        - email
    NewUser:
      description: "schema for defining a new user with basic information such as email and password."
      additionalProperties: false
      allOf:
        - $ref: "#/components/schemas/UserBase"
      properties:
        password:
          type: string
          minLength: 6
          maxLength: 20
          description: "user password"
      required:
        - password
    User:
      description: "schema representing a user with basic details and a self-reference link."
      additionalProperties: false
      allOf:
        - $ref: "#/components/schemas/UserBase"
      properties:
        id:
          type: integer
          description: "unique identifier for the user"
      # Links according to HATEOAS principle
        self:
          type: string
          format: uri
          description: "Link to the current user"
      required:
        - id
    Users:
      description: "schema of a collection of user objects."
      type: object
      additionalProperties: false
      properties:
        $schema:
          type: string
          description: "path to the JSON Schema to use for the validation of the JSON data structure describing the users"
        totalPages:
          type: integer
          minimum: 0
          description: "total number of pages"
        currentPage:
          type: integer
          minimum: 0
          description: "the page included in this object"
        totalItems:
          type: integer
          minimum: 0
          description: "the total number of items in the current page"
        users:
          type: array
          items:
            $ref: "#/components/schemas/User"
          description: "the films of the current page"
        # Links according to HATEOAS principle
        next:
          type: string
          format: uri
          description: "Link to the next page of users (if available)"
        previous:
          type: string
          format: uri
          description: "Link to the previous page of users (if available)"
      required:
        - totalPages
        - users
...
    Image:
      description: "schema that represents an image of a public film"
      type: object
      additionalProperties: false
      properties:
        $schema:
          type: string
          description: "path to the JSON Schema to use for the validation of the JSON data structure describing the image"
        id:
          type: integer
          description: "The unique identifier for the image."
        originalname:
          type: string
          description: "The original name of the image file, excluding the file extension."
        filename:
          type: string
          description: "The generated filename for the uploaded image, typically including the file extension."
        filmId:
          type: integer
          description: "The unique identifier of the film associated with the image."
        # Links according to HATEOAS principle
        self:
          type: string
          format: uri
          description: "Link to the current image resource"
        delete:
          type: string
          format: uri
          description: "Link to delete the image resource"
      required:
        - id
        - name
        - type
    Images:
      description: "schema for Images data structures"
      additionalProperties: false
      type: array
      items:
        $ref: '#/components/schemas/Image'
    ErrorResponse:
      description: "Represents an error response returned by the API."
      type: object
      additionalProperties: false
      properties:
        $schema:
          type: string
          description: "path to the JSON Schema used for validating the error response structure"
        code:
          type: integer
          description: "error code"
        message:
          type: string
          description: "a brief description of the error"
      required:
        - code
        - message
  securitySchemes:
    cookieAuth:
      description: "JWT to identify the authenticated user."
      type: apiKey
      in: cookie
      name: connect.sid